Refactor bot to a modular structure with Docker support
This commit refactors the entire bot from a single-file script into a scalable, modular architecture. The goal was to improve the project's hierarchy and database integration, making it more maintainable and easier to extend.

Key changes include:
- **Modular Structure:** The application logic is now split into separate directories for handlers, keyboards, states, middlewares, and core components (database, config).
- **Configuration:** Replaced `python-dotenv` with `pydantic-settings` for robust environment variable management.
- **Containerization:** Added a `Dockerfile` and `docker-compose.yml` to allow the bot to be run in a containerized environment, simplifying deployment and setup.
- **Database Abstraction:** The existing database logic has been moved into its own `Database` class in `app/core/database.py`.
- **Updated Documentation:** The `README.md` has been rewritten to reflect the new project structure and includes instructions for running the bot with Docker.
