+++ README.md
@@ -1,39 +1,73 @@
 # Pawer Telegram Bot
 
-Это простой Telegram-бот, созданный с использованием библиотеки `aiogram`.
+Это Telegram-бот, созданный с использованием `aiogram` 3.x, с модульной структурой и поддержкой Docker.
 
-## Установка
+## 🚀 Особенности
 
-1.  Клонируйте репозиторий:
-    ```bash
-    git clone <URL репозитория>
-    cd <название папки>
-    ```
+-   **Модульная структура**: Логика разделена на хендлеры, клавиатуры, состояния и т.д.
+-   **Docker-Ready**: Полностью настроен для запуска в Docker-контейнере.
+-   **Конфигурация через `.env`**: Удобное управление настройками с помощью `pydantic-settings`.
+-   **База данных**: Использует `aiosqlite` для асинхронной работы с базой данных SQLite.
 
-2.  Создайте и активируйте виртуальное окружение (рекомендуется):
-    ```bash
-    python -m venv venv
-    source venv/bin/activate  # Для Windows: venv\Scripts\activate
-    ```
+## 📂 Структура проекта
 
-3.  Установите зависимости:
-    ```bash
-    pip install -r requirements.txt
-    ```
-
-## Настройка
-
-1.  Получите токен для вашего бота у [@BotFather](https://t.me/BotFather) в Telegram.
-
-2.  Создайте в корне проекта файл `.env` и добавьте в него ваш токен:
-    ```
-    TELEGRAM_BOT_TOKEN="1234567890:ABCDEFGHIJKLMNOPQRSTUVWXYZ"
-    ```
-    Либо установите переменную окружения `TELEGRAM_BOT_TOKEN` любым другим удобным для вас способом.
-
-## Запуск
-
-Для запуска бота выполните команду:
-```bash
-python bot.py
 ```
+.
+├── bot/
+│   ├── app/
+│   │   ├── core/         # Ядро приложения (конфигурация, БД)
+│   │   ├── handlers/     # Обработчики сообщений и колбеков
+│   │   ├── keyboards/    # Клавиатуры
+│   │   ├── middlewares/  # Middleware
+│   │   └── states/       # FSM состояния
+│   ├── .env            # Файл с переменными окружения
+│   ├── Dockerfile      # Dockerfile для сборки образа
+│   ├── requirements.txt# Зависимости
+│   └── run.py          # Точка входа для запуска бота
+├── docker-compose.yml  # Файл для управления Docker-контейнером
+└── README.md
+```
+
+## ⚙️ Настройка и запуск
+
+### 1. Клонирование репозитория
+
+```bash
+git clone <URL репозитория>
+cd <название папки>
+```
+
+### 2. Настройка окружения
+
+1.  Перейдите в директорию `bot`.
+2.  Создайте файл `.env` на основе примера `.env.example` (если он есть) или создайте его с нуля.
+3.  Откройте файл `bot/.env` и укажите ваш токен:
+
+    ```env
+    BOT_TOKEN="1234567890:ABCDEFGHIJKLMNOPQRSTUVWXYZ"
+    DATABASE_PATH="bot_database.db"
+    LOG_LEVEL="INFO"
+    ```
+
+    -   `BOT_TOKEN`: Токен, полученный от [@BotFather](https://t.me/BotFather).
+    -   `DATABASE_PATH`: Путь к файлу базы данных внутри контейнера.
+    -   `LOG_LEVEL`: Уровень логирования (DEBUG, INFO, WARNING, ERROR).
+
+### 3. Запуск с помощью Docker
+
+Убедитесь, что у вас установлен Docker и Docker Compose.
+
+Выполните команду из корневой директории проекта:
+
+```bash
+docker-compose up --build
+```
+
+Бот будет запущен в фоновом режиме. Чтобы остановить его, выполните:
+
+```bash
+docker-compose down
+```
+
+---
+*Этот README был обновлен для отражения новой структуры проекта.*
