+++ bot/.env
@@ -1,0 +1,8 @@
+# Telegram Bot Token from @BotFather
+BOT_TOKEN="1234567890:ABCDEFGHIJKLMNOPQRSTUVWXYZ"
+
+# Path to the database file
+DATABASE_PATH="bot_database.db"
+
+# Logging level (e.g., INFO, DEBUG, WARNING)
+LOG_LEVEL="INFO"
