+++ bot/app/core/__init__.py
@@ -1,0 +1,4 @@
+from .config import settings
+from .database import Database
+
+__all__ = ["settings", "Database"]
