+++ bot/app/core/config.py
@@ -1,0 +1,12 @@
+from pydantic_settings import BaseSettings, SettingsConfigDict
+
+
+class Settings(BaseSettings):
+    model_config = SettingsConfigDict(env_file=".env", env_file_encoding="utf-8")
+
+    BOT_TOKEN: str
+    DATABASE_PATH: str
+    LOG_LEVEL: str = "INFO"
+
+
+settings = Settings()
