+++ bot/app/keyboards/__init__.py
@@ -1,0 +1,7 @@
+from .inline import get_main_keyboard, get_settings_keyboard, get_back_keyboard
+
+__all__ = [
+    "get_main_keyboard",
+    "get_settings_keyboard",
+    "get_back_keyboard",
+]
