+++ bot/app/middlewares/__init__.py
@@ -1,0 +1,3 @@
+from .user_middleware import user_middleware
+
+__all__ = ["user_middleware"]
