+++ bot/app/states/__init__.py
@@ -1,0 +1,3 @@
+from .user_states import UserStates
+
+__all__ = ["UserStates"]
