+++ bot/app/states/user_states.py
@@ -1,0 +1,6 @@
+from aiogram.fsm.state import State, StatesGroup
+
+
+class UserStates(StatesGroup):
+    waiting_for_name = State()
+    waiting_for_feedback = State()
