+++ bot/requirements.txt
@@ -1,3 +1,3 @@
 aiogram==3.4.1
 aiosqlite==0.19.0
-python-dotenv==1.0.0
+pydantic-settings==2.2.1
