+++ docker-compose.yml
@@ -1,0 +1,13 @@
+version: '3.8'
+
+services:
+  bot:
+    build:
+      context: ./bot
+      dockerfile: Dockerfile
+    container_name: telegram_bot
+    env_file:
+      - ./bot/.env
+    volumes:
+      - ./bot/bot_database.db:/usr/src/app/bot_database.db
+    restart: unless-stopped
